requirements.txt
streamlit==1.24.1
pandas==2.0.3
langchain==0.0.288
openai==0.27.8
python-environ==0.4.54
tabulate==0.9.0